extracted media-selection-overlay into seperate container and added readme
