+++ src/Sulu/Bundle/MediaBundle/Resources/js/containers/MediaSelectionOverlay/README.md
@@ -1,0 +1,5 @@
+This component uses a [`MediaCollection`](#mediacollection) in an [`Overlay`](#overlay) to offer the possibility to select a medias. 
+The `onConfirm` callback is called with the media objects as soon as the confirm button is clicked.
+
+This component mainly serves as base for the [`SingleMediaSelectionOverlay`](#singlemediaselectionoverlay) and
+[`MultiMediaSelectionOverlay`](#multimediaselectionoverlay) components.
