+++ src/Sulu/Bundle/MediaBundle/Resources/js/containers/MediaSelectionOverlay/index.js
@@ -1,0 +1,4 @@
+// @flow
+import MediaSelectionOverlay from './MediaSelectionOverlay';
+
+export default MediaSelectionOverlay;
