+++ src/Sulu/Bundle/MediaBundle/Resources/js/containers/MediaSelectionOverlay/mediaSelectionOverlay.scss
